services/monitoring/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-monitoring"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "prometheus-client>=0.19.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
